improvement: Rename core-nats to core-minimal module, add MQTT as protocol
