+++ .github/workflows/build.yml
@@ -70,13 +70,13 @@ jobs:
           push: true
           tags: ${{ env.DOCKERHUB_APACHE_REPO }}/backend:${{ env.MVN_VERSION }}
 
-      - name: Build and Push Docker Backend Nats Image
+      - name: Build and Push Docker Core Minimal Image
         uses: docker/build-push-action@v5
         with:
-          context: ./streampipes-service-core-nats
+          context: ./streampipes-service-core-minimal
           platforms: linux/amd64,linux/arm64/v8,linux/arm/v7
           push: true
-          tags: ${{ env.DOCKERHUB_APACHE_REPO }}/backend-nats:${{ env.MVN_VERSION }}
+          tags: ${{ env.DOCKERHUB_APACHE_REPO }}/backend-minimal:${{ env.MVN_VERSION }}
 
       - name: Build UI
         working-directory: ./ui
