+++ streampipes-extensions/streampipes-extensions-iiot-minimal/pom.xml
@@ -63,6 +63,12 @@
 
         <dependency>
             <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-messaging-mqtt</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-processors-filters-jvm</artifactId>
             <version>0.93.0-SNAPSHOT</version>
             <exclusions>
