+++ streampipes-service-core-minimal/pom.xml
@@ -27,7 +27,7 @@
         <version>0.93.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>streampipes-service-core-nats</artifactId>
+    <artifactId>streampipes-service-core-minimal</artifactId>
 
     <dependencies>
         <dependency>
@@ -42,10 +42,6 @@
                 </exclusion>
                 <exclusion>
                     <groupId>org.apache.streampipes</groupId>
-                    <artifactId>streampipes-messaging-mqtt</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>org.apache.streampipes</groupId>
                     <artifactId>streampipes-messaging-kafka</artifactId>
                 </exclusion>
                 <exclusion>
@@ -53,6 +49,11 @@
                     <artifactId>streampipes-messaging-pulsar</artifactId>
                 </exclusion>
             </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-messaging-mqtt</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
 
@@ -67,7 +68,7 @@
                             <goal>repackage</goal>
                         </goals>
                         <configuration>
-                            <mainClass>org.apache.streampipes.service.core.nats.StreamPipesCoreApplicationNats
+                            <mainClass>org.apache.streampipes.service.core.minimal.StreamPipesCoreApplicationMinimal
                             </mainClass>
                         </configuration>
                     </execution>
@@ -88,7 +89,7 @@
                 <artifactId>maven-checkstyle-plugin</artifactId>
             </plugin>
         </plugins>
-        <finalName>streampipes-core-nats</finalName>
+        <finalName>streampipes-core-minimal</finalName>
     </build>
 
 </project>
